+++ sdk/monitor/Azure.Monitor.OpenTelemetry/tests/Azure.Monitor.OpenTelemetry.Demo/ProgramNet461Compat.cs
@@ -2,16 +2,15 @@
 // Licensed under the MIT License.
 
 #if NET461
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore;
 
 namespace Azure.Monitor.OpenTelemetry.Demo
 {
     public partial class Program
     {
-        public static void Main(string[] args) => CreateWebHostBuilder(args).Build().Run();
-
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) => WebHost.CreateDefaultBuilder(args).UseStartup<Startup>();
+        public static void Main(string[] args)
+        {
+        }
     }
 }
+
 #endif
